Move FixedProps to module's typings
