+++ tsconfig.json
@@ -28,6 +28,7 @@
   "include": [
     "./src/**/*.js",
     "./src/**/*.ts",
-    "./src/**/*.d.ts"
+    "./src/**/*.d.ts",
+    "./typings/**/*.d.ts"
   ]
 }
